pyproject.toml
[tool.poetry]
name = "keyrunes-python-sdk"
version = "0.1.0"
description = "Python SDK for Keyrunes Authorization System"
authors = ["keyrunes <contact@singularjourney.host>"]
maintainers = ["jonatasoli <contact@jonatasoli.dev>"]
readme = "README.md"
license = "MIT"
homepage = "https://keyrunes.com"
repository = "https://github.com/jonatasoli/keyrunes-python-sdk"
documentation = "https://github.com/jonatasoli/keyrunes-python-sdk#readme"
keywords = ["keyrunes", "authorization", "rbac", "abac", "security", "authentication", "permissions"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security",
    "Topic :: System :: Systems Administration :: Authentication/Directory",
]
packages = [{include = "keyrunes_sdk"}]
include = ["LICENSE", "README.md", "CHANGELOG.md"]
exclude = ["tests", "examples", "docs", "*.pyc", "__pycache__"]

[tool.poetry.dependencies]
python = ">=3.10.1,<4.0"
pydantic = {extras = ["email"], version = "^2.0.0"}
pyjwt = "^2.9.0"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
factory-boy = "^3.3.0"
faker = "^19.0.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
taskipy = "^1.12.0"
towncrier = "^25.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=keyrunes_sdk",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "-v",
]

[tool.black]
line-length = 100
target-version = ['py314']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.taskipy.tasks]
test = "pytest"
cov = "pytest --cov=keyrunes_sdk --cov-report=term-missing --cov-report=xml"
lint = "flake8 keyrunes_sdk tests"
format = "black keyrunes_sdk tests examples && isort keyrunes_sdk tests examples"
type-check = "mypy keyrunes_sdk"
clean = "rm -rf .pytest_cache htmlcov .coverage coverage.xml dist/ build/ *.egg-info && find . -type d -name __pycache__ -exec rm -rf {} + && find . -type f -name '*.pyc' -delete"
docker-up = "docker-compose up -d"
docker-down = "docker-compose down"
docker-logs = "docker-compose logs -f keyrunes"
docker-ps = "docker-compose ps"
docker-restart = "docker-compose restart"
docker-clean = "docker-compose down -v && docker system prune -f"
test-local = "docker-compose up -d && sleep 10 && poetry run python examples/test_local.py"
example-basic = "docker-compose up -d && sleep 10 && poetry run python examples/basic_usage.py"
example-global = "docker-compose up -d && sleep 10 && poetry run python examples/global_client_usage.py"
build = "poetry build"
check-package = "poetry build && twine check dist/*"
publish-test = "poetry build && twine check dist/* && poetry publish -r test-pypi"
publish = "poetry build && twine check dist/* && poetry publish"
version-patch = "poetry version patch"
version-minor = "poetry version minor"
version-major = "poetry version major"
all = "poetry run task format && poetry run task lint && poetry run task type-check && poetry run task test"
ci = "poetry run task lint && poetry run task type-check && poetry run task test"
